#8725-R: move VectorSimplifyMethod to other file
