+++ src/core/qgsvectorlayer.cpp
@@ -4217,24 +4217,3 @@ bool QgsAttributeEditorRelation::init( QgsRelationManager* relationManager )
   mRelation = relationManager->relation( mRelationId );
   return mRelation.isValid();
 }
-
-
-QgsVectorSimplifyMethod::QgsVectorSimplifyMethod()
-  : mSimplifyHints( QGis::DEFAULT_MAPTOPIXEL_THRESHOLD > 1 ? QgsVectorLayer::FullSimplification : QgsVectorLayer::GeometrySimplification )
-  , mThreshold( QGis::DEFAULT_MAPTOPIXEL_THRESHOLD )
-  , mLocalOptimization( true )
-{
-}
-
-QgsVectorSimplifyMethod::QgsVectorSimplifyMethod( const QgsVectorSimplifyMethod &rh )
-{
-  operator=( rh );
-}
-
-QgsVectorSimplifyMethod& QgsVectorSimplifyMethod::operator=( const QgsVectorSimplifyMethod &rh )
-{
-  mSimplifyHints = rh.mSimplifyHints;
-  mThreshold = rh.mThreshold;
-  mLocalOptimization = rh.mLocalOptimization;
-  return *this;
-}
