+++ src/core/qgsvectorlayer.h
@@ -32,6 +32,7 @@
 #include "qgssnapper.h"
 #include "qgsfield.h"
 #include "qgsrelation.h"
+#include "qgsvectorsimplifymethod.h"
 
 class QPainter;
 class QImage;
@@ -177,41 +178,6 @@ struct CORE_EXPORT QgsVectorJoinInfo
   int targetFieldIndex;
   /**Join field index in the source layer. For backward compatibility with 1.x (x>=7)*/
   int joinFieldIndex;
-};
-
-/** This class contains information how to simplify geometries fetched from a vector layer */
-class CORE_EXPORT QgsVectorSimplifyMethod
-{
-  public:
-    //! construct a default object
-    QgsVectorSimplifyMethod();
-    //! copy constructor
-    QgsVectorSimplifyMethod( const QgsVectorSimplifyMethod& rh );
-    //! assignment operator
-    QgsVectorSimplifyMethod& operator=( const QgsVectorSimplifyMethod& rh );
-
-    /** Sets the simplification hints of the vector layer managed */
-    void setSimplifyHints( int simplifyHints ) { mSimplifyHints = simplifyHints; }
-    /** Gets the simplification hints of the vector layer managed */
-    inline int simplifyHints() const { return mSimplifyHints; }
-
-    /** Sets the simplification threshold of the vector layer managed */
-    void setThreshold( float threshold ) { mThreshold = threshold; }
-    /** Gets the simplification threshold of the vector layer managed */
-    inline float threshold() const { return mThreshold; }
-
-    /** Sets where the simplification executes, after fetch the geometries from provider, or when supported, in provider before fetch the geometries */
-    void setForceLocalOptimization( bool localOptimization ) { mLocalOptimization = localOptimization; }
-    /** Gets where the simplification executes, after fetch the geometries from provider, or when supported, in provider before fetch the geometries */
-    inline bool forceLocalOptimization() const { return mLocalOptimization; }
-
-  private:
-    /** Simplification hints for fast rendering of features of the vector layer managed */
-    int mSimplifyHints;
-    /** Simplification threshold */
-    float mThreshold;
-    /** Simplification executes after fetch the geometries from provider, otherwise it executes, when supported, in provider before fetch the geometries */
-    bool mLocalOptimization;
 };
 
 /** \ingroup core
